Switch Branch to Option. Add Rust parsing node type. Extract BulletParser to own file
