+++ src-tauri/src/mdt/mod.rs
@@ -3,3 +3,6 @@ include!("cmds.rs");
 include!("file_parse.rs");
 include!("file_write.rs");
 include!("structs.rs");
+
+// - Parsers
+include!("parsers/bullet_file_parser.rs");
