+++ src-tauri/test/data/03_basic_encoding.md
@@ -1,0 +1,16 @@
+# Bullets (md-decision-trees)
+
+D: What to do about X?
+* O: Could do Y
+  * P: Would help now
+  * C: Restricts interface
+* O: Could do Z
+  * P: Would help later
+  * C: Takes longer to implement
+    * N: Not a big deal - have plenty of time this sprint
+  * C: Requires more user knowledge
+
+D: What about the other thing?
+* O: This way
+* O: That way
+* N: Need to have meeting about this
